# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Agriculture API Docs",
        "description": "API Docs for Agriculture Reports",
        "contact": {
            "email": "sajeeshe@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/tractors": {
            "get": {
                "tags": [
                    "Tractors"
                ],
                "summary": "List All Tractors",
                "responses": {
                    "200": {
                        "description": "Return Tractors list"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/tractor/add": {
            "post": {
                "tags": [
                    "Tractors"
                ],
                "summary": "Add new Tractor",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Tractor name",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return Created Tractor"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/tractor/{slug}": {
            "get": {
                "tags": [
                    "Tractors"
                ],
                "summary": "Display individual Tractor detail",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Tractor Slug",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return selected Tractor details"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/tractor/update/{slug}": {
            "put": {
                "tags": [
                    "Tractors"
                ],
                "summary": "Add new Tractor",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Tractor Name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Tractor Slug",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return Created Tractor"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/tractor/delete/{slug}": {
            "delete": {
                "tags": [
                    "Tractors"
                ],
                "summary": "Delete a Tractor",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Tractor Slug",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return Deleted Tractor details"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/user/list": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "List All Users",
                "responses": {
                    "200": {
                        "description": "Return Users list"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/user/add": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Add new User ( Admin can only add new user )",
                "parameters": [
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "First name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "Last name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Role ( User | Supervisor | Admin )",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return Created Field"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/user/view/{slug}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Display single user details ",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "User Slug",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return selected User details"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/user/update/{slug}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Add new User ( Admin can only add new user )",
                "parameters": [
                    {
                        "name": "first_name",
                        "in": "query",
                        "description": "First name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "last_name",
                        "in": "query",
                        "description": "Last name",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Email",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "query",
                        "description": "Password",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "role",
                        "in": "query",
                        "description": "Role ( User | Supervisor | Admin )",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "User Slug",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return Updated User"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        },
        "/api/user/delete/{slug}": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "Delete selected user ",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "User Slug",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Return selected User details"
                    }
                },
                "security": [
                    {
                        "token": []
                    }
                ]
            }
        }
    },
...
    "securityDefinitions": {
        "token": {
            "type": "apiKey",
            "name": "token",
            "in": "query"
        }
    }
}